processor/requirements.txt
azure-functions>=1.19.0
azure-storage-blob>=12.20.0
azure-search-documents>=11.4.0
azure-eventgrid>=4.17.0
azure-identity>=1.15.0
azure-ai-inference>=1.0.0b1

openai>=1.30.0
pymupdf4llm>=0.0.6
pymupdf>=1.23.0

pydantic>=2.7.0
python-dotenv>=1.0.0
tenacity>=8.3.0

# For development
pytest>=8.2.0
pytest-asyncio>=0.23.0
